Found a problem.
The rabbit moves deeper into the land modules along the x axis so much so that he is behind all of the objects that he is supposed to be jumping over. This only happens after playing the game for a couple of minutes. I am in the process of finding what is causing the problem so I can fix.it.