Audio and VFX
I added in audio to the star so that when the rabbit hit it, the star would ding. I then spent the rest of my time trying to figure out how to add a flash to the star on collision. I wasn't very successful in this (there were a lot of errors that I could spend a whole other afternoon-night on) and so I ended up putting a light tail in behind coyote to get the VFX quota.